+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 4.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Fallback to default event field value in accessor, if location reference is not set.
+  [cekk]
 
 
 4.0 (2019-06-06)
@@ -33,7 +34,6 @@ New features:
 
     - Customize event_summary view and portlet event renderer to properly show location info.
       [cekk]
-    - Fallback to default event field value in accessor, if location reference is not set.
 
 Bug fixes: